move function to performance_manager
